requirements.txt
altair==5.1.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
deta==1.2.0
extra-streamlit-components==0.1.60
gitdb==4.0.11
GitPython==3.1.40
idna==3.5
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.22
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
zipp==3.17.0
